A complete set of diagnostic test programs for the IBM 1620 Jr.
Most of these programs are the original IBM 1620 diagnostics
with the exception of CS01 and FP01 which were written for the
IBM 1620 Jr.

All of these programs are built into the simulator and can be
loaded and run from the front panel following the directions
in the IBM 1620 Jr Diagnostic Manual.